+++ spring-batch-db-cluster-core/src/main/java/io/github/jchejarla/springbatch/clustering/partition/PartitioningMode.java
@@ -16,7 +16,7 @@
 package io.github.jchejarla.springbatch.clustering.partition;
 
 public enum PartitioningMode {
-    SCALE_UP,
+    ROUND_ROBIN,
     FIXED_NODE_COUNT,
-    ROUND_ROBIN
+    LEAST_LOADED
 }